@@ -1,5 +1,5 @@
-import { Component, HostListener, OnInit } from '@angular/core';
-import { FormArray, FormBuilder, FormGroup } from '@angular/forms';
+import { Component, ElementRef, HostListener, OnInit, ViewChild } from '@angular/core';
+import { UntypedFormArray, UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { Area } from 'app/models/area';
 import { AreaService } from 'app/services/area.service';
 import { MesaEstadoService } from 'app/services/mesa-estado.service';
@@ -16,7 +16,7 @@ import { environment } from 'environments/environment';
 
 export class MesaComponent implements OnInit {
 
-  itemForms: FormArray = this.fb.array([]);
+  itemForms: UntypedFormArray = this.fb.array([]);
   userList: [];
   stateList: [];
   shapeList: [];
@@ -28,7 +28,7 @@ export class MesaComponent implements OnInit {
   factMesas = 0;
 
   constructor(
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private mesaService: MesaService,
     private userService: UserService,
     private mesaFormaService: MesaFormaService,
@@ -99,7 +99,7 @@ export class MesaComponent implements OnInit {
     }));
   }
 
-  recordSubmit(fg: FormGroup) {
+  recordSubmit(fg: UntypedFormGroup) {
     if (fg.value.id == 0)
       this.mesaService.postMesa(fg.value).subscribe(
         (res: any) => {
@@ -156,16 +156,26 @@ export class MesaComponent implements OnInit {
   
   @HostListener('window:resize', ['$event'])
     onResize(event) {
+      console.log(this.mesitas.nativeElement.offsetWidth);
+      console.log(this.mesitas.nativeElement.offsetHeight);
       let clientWidth = document.getElementById('mesas').clientWidth;
       this.factMesas = clientWidth / this.areaSelectedLength;
       console.log(document.getElementById('mesas'));
       console.log(document.getElementById('mesas').clientWidth);
     }
 
-
-  /*ngAfterViewChecked() {
-    let clientWidth = document.getElementById('mesas').clientWidth;
-    this.factMesas = clientWidth / this.areaSelectedLength;    
-  }*/
+//Master_rama
+  @ViewChild('mesitas')
+  mesitas: ElementRef;
+  
+  
+  //Facundo_rama
+    ngAfterViewChecked() {
+      if (document.getElementById('mesas')) {
+        let clientWidth = document.getElementById('mesas').clientWidth;
+        this.factMesas = clientWidth / this.areaSelectedLength;        
+      }
+  }
+  
 
 }
