@@ -1,4 +1,5 @@
-import { Component} from '@angular/core';
+import { Component, OnDestroy, OnInit} from '@angular/core';
+import { SignalrService } from './services/signalr.service';
 
 
 @Component({
@@ -6,6 +7,22 @@ import { Component} from '@angular/core';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
+export class AppComponent implements OnInit, OnDestroy {
+  constructor(
+    private signalrService: SignalrService
+  ) {}
+
+  ngOnInit() {
+     this.signalrService.startConnection();
+
+     setTimeout(() => {
+       this.signalrService.askServerListener();
+     }, 2000);
+  }
+
+  ngOnDestroy() {
+    this.signalrService.hubConnection.off('askServerResponse');
+    this.signalrService.hubConnection.stop();
+  }
 
 }
