@@ -1,42 +1,38 @@
 import { Component, OnInit } from '@angular/core';
-declare var $: any;
+import { NotificationService } from 'app/services/notification.service';
+import { SignalrService } from 'app/services/signalr.service';
+import { Notification } from 'app/models/notification';
+
 @Component({
   selector: 'app-notifications',
   templateUrl: './notifications.component.html',
   styleUrls: ['./notifications.component.css']
 })
 export class NotificationsComponent implements OnInit {
+  notifications: Notification[];
 
-  constructor() { }
-  showNotification(from, align){
-      const type = ['','info','success','warning','danger'];
+  constructor(
+    private notificationService: NotificationService,
+    private signalrService: SignalrService
+  ) { }
 
-      const color = Math.floor((Math.random() * 4) + 1);
 
-      $.notify({
-          icon: "notifications",
-          message: "Welcome to <b>Material Dashboard</b> - a beautiful freebie for every web developer."
-
-      },{
-          type: type[color],
-          timer: 4000,
-          placement: {
-              from: from,
-              align: align
-          },
-          template: '<div data-notify="container" class="col-xl-4 col-lg-4 col-11 col-sm-4 col-md-4 alert alert-{0} alert-with-icon" role="alert">' +
-            '<button mat-button  type="button" aria-hidden="true" class="close mat-button" data-notify="dismiss">  <i class="material-icons">close</i></button>' +
-            '<i class="material-icons" data-notify="icon">notifications</i> ' +
-            '<span data-notify="title">{1}</span> ' +
-            '<span data-notify="message">{2}</span>' +
-            '<div class="progress" data-notify="progressbar">' +
-              '<div class="progress-bar progress-bar-{0}" role="progressbar" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100" style="width: 0%;"></div>' +
-            '</div>' +
-            '<a href="{3}" target="{4}" data-notify="url"></a>' +
-          '</div>'
-      });
-  }
   ngOnInit() {
-  }
+    this.notificationService.getNotifications().subscribe(notifications => {
 
+    // ordenando notificaciones por fecha de creación de forma descendente
+    const sortedDesc = notifications.sort(
+      (objA: Notification, objB: Notification) => new Date(objB.date).getTime() - new Date(objA.date).getTime()
+      );
+    
+    this.notifications = sortedDesc as Notification[];
+    });
+  }
+  deleteNotification(notification: Notification) {
+    this.notificationService.deleteNotificationFromStorage(notification);
+  }
+  setNotificationReaded(notification: Notification) {
+    notification.readed = true;
+    this.notificationService.updateNotification(notification);
+  }
 }
